requirements/local.txt
-r base.txt

django-ptpython==1.0.2
